fix(Workspace): return correct int value

+++ src/QUI/Workspace/Manager.php
@@ -142,7 +142,7 @@ class Manager
             'minWidth' => $minWidth
         ]);
 
-        return QUI::getDataBase()->getPDO()->lastInsertId('id');
+        return (int)QUI::getDataBase()->getPDO()->lastInsertId('id');
     }
 
     /**
